Moved filters to the separate file
