+++ filters.py
@@ -1,0 +1,46 @@
+import operator
+from functools import reduce
+from telegram import Message
+from telegram.ext import BaseFilter, Filters, MessageFilter
+from urlextract import URLExtract
+
+
+class ChatMessageOnly(MessageFilter):
+    def filter(self, message: Message) -> bool:
+        return message.forward_from_message_id is None
+
+
+def with_default_filters(*filters: BaseFilter) -> BaseFilter:
+    """Apply default filters to the given filter classes"""
+    default_filters = [
+        Filters.text,
+        ChatMessageOnly(),
+    ]
+    return reduce(operator.and_, [*default_filters, *filters])  # МАМА Я УМЕЮ ФУНКЦИОНАЛЬНО ПРОГРАММИРОВАТЬ
+
+
+class IsMessageOnBehalfOfChat(MessageFilter):
+    def filter(self, message: Message) -> bool:
+        return message.sender_chat is not None
+
+
+class ContainsTelegramContact(MessageFilter):
+    def filter(self, message: Message) -> bool:
+        return ' @' in message.text
+
+
+class ContainsLink(MessageFilter):
+    def __init__(self) -> None:
+        self.extractor = URLExtract()
+
+    def filter(self, message: Message) -> bool:
+        return len(self.extractor.find_urls(message.text)) >= 1
+
+
+__all__ = [
+    'ContainsTelegramContact',
+    'ContainsLink',
+    'ChatMessageOnly',
+    'IsMessageOnBehalfOfChat',
+    'with_default_filters',
+]
